Shift click on a tag, highlights all states matching that tag
If, for example, you are tagging your states with the request id, otel
trace ID, span ID, etc, you may want to see all the work that is
coming from the same request.

By default, clicking on a tag only shows the same states that contain
the tag. This change allows you to hold down shift to see all the
states which match the tag.